make Multiaddr a newtype of []bytes
This allows us to send multiaddrs on the wire without copying (through the
`Bytes()` method). It also avoids one allocation/indirection per multiaddr.

This is an alternative (or stop-gap) to #62 as #62 is invasive. Unlike #62, this
PR doesn't break `Multiaddr == nil` or `maddr1.Equals(maddr2)`. However,
it *doesn't* allow multiaddrs to be used as map keys :(.