fix: update /x/exp package and fix sorting
Update /x/exp to match the go 1.21 "slices" package.

The signature of slices.SortFunc changed to now require the function return an
integer instead of a boolean for sorting. Fix the call to slices.SortFunc.

This fixes the build on Go 1.21 when downstream projects import a newer version
of /x/exp than the current version of multiaddr.

Signed-off-by: Christian Stewart <christian@aperture.us>